Update install.sh
Removed Root from running the service:
A service account named minecraft is created with adduser.
The Minecraft server directory (/opt/minecraft) is created and its ownership is set to the minecraft user. The systemd service file is updated to run the server as the minecraft user. Permissions for the Minecraft server directory are set to allow the minecraft user to read, write, and execute, while the group has read and execute permissions.